requirements.txt
SpeechRecognition
gTTS
pyaudio
pygame
Flask
google-cloud-speech
PyAudio
pydub
openai-whisper
